fix: Build by moving from findbugs to spotbugs and removing Java 7 tests
Actually the Travis build on master is broken because of an issue with findbugs 3.
Since findbugs is deprecated, its seems to be better to switch to spotbugs, the successor.
The pom.xml has been updated accordingly and some low hanging fruits have been fixed.

However the test show some concurrency issue, which might be considered to be fixed.

The test for Java 7 has been switched off in Travis as (a) the build is broken (see one of the latest Travis logs) and (b) spotbugs requires Java 8+