Read all data available before returning to select
There appears to be some buffering interaction with the native
`poll` call here that causes unread bytes "on the wire" to be left
"in limbo" if they are not read before a subsequent POLLIN has
fired. As a result, there are too few POLLIN events to read all
the data off the wire and the server eventually hangs.

A client attempting to write and then read data from the server
also eventually hangs, because it has written all it can write but
not received enough data back.

More research is needed to know whether this is an expected
interaction between `poll` and the buffering for sockets (or more
specifically unix domain sockets), possibly specific to Darwin.

See #82